split up geo package into separate files, fix executable
