+++ geo/grid_test.go
@@ -36,100 +36,6 @@ var path3 = []GeoPoint{
 	GeoPoint{42.364667, -71.102567},
 }
 	
-func TestPathBoundary(t *testing.T) {
-
-	// find the min / max lat long for one path
-	boundary1 := PathBoundary(path1)
-
-	// boundary should be:
-	// {
-	//   maxLat: 42.365592
-	//   minLat: 42.361439
-	//   maxLon: -71.103875
-	//   minLon: -71.116022
-	// }
-
-	t1 := GeoRect{
-		NorthWest: GeoPoint{42.365592, -71.116022},
-		SouthEast: GeoPoint{42.361439, -71.103875},
-	}
-	if !reflect.DeepEqual(boundary1, t1) {
-		t.Errorf("incorrect boundary structure for 1-path example")
-	}
-
-	// then for multiple
-	boundary2 := PathBoundary(path1, path2, path3)
-
-	// boundary should be
-	// {
-	//   maxLat: 42.365592
-	//   minLat: 42.359000
-	//   maxLon: -71.096162
-	//   minLon: -71.116022
-	// }
-
-	t2 := GeoRect{
-		NorthWest: GeoPoint{42.365592, -71.116022},
-		SouthEast: GeoPoint{42.359000, -71.096162},
-	}
-
-	if !reflect.DeepEqual(boundary2, t2) {
-		t.Errorf("incorrect boundary structure for 3-path example")
-	}
-}
-
-func TestGeoDistance(t *testing.T) {
-
-	// two points relatively close together (half a mile)
-	d1 := GeoDistance(GeoPoint{42.353381, -71.107131}, GeoPoint{42.356941, -71.092647})
-	if math.Abs(d1 - 1250) > 7 {
-		t.Errorf("GeoDistance off for two close points")
-	}
-
-	// two points very close together (a few feet)
-	d2 := GeoDistance(GeoPoint{42.364378, -71.114549}, GeoPoint{42.364447, -71.114603})
-	if math.Abs(d2 - 10) > 7 {
-		t.Errorf("GeoDistance off for two very close points")
-	}
-	
-	// the same two points
-	d3 := GeoDistance(GeoPoint{42.365673, -71.104100}, GeoPoint{42.365673, -71.104100})
-	if d3 != 0 {
-		t.Errorf("GeoDistance off for exact two points")
-	}
-	
-	// two points at the same longitude
-	d4 := GeoDistance(GeoPoint{42.365673, -71.104100}, GeoPoint{42.845683, -71.104100})
-	if math.Abs(d4 - 53370) > 7 {
-		t.Errorf("GeoDistance off for two points on same line of longitude")
-	}
-	
-	// two points at the same latitude
-	d5 := GeoDistance(GeoPoint{42.365673, -71.304331}, GeoPoint{42.365673, -71.104100})
-	if math.Abs(d5 - 16450) > 7 {
-		t.Errorf("GeoDistance off for two points on same line of latitude")
-	}
-	
-	// two points relatively far away with varying lat / lon (~10 kilometers)
-	d6 := GeoDistance(GeoPoint{42.365121, -71.212806}, GeoPoint{42.366810, -71.068591})
-	if math.Abs(d6 - 11850) > 7 {
-		t.Errorf("GeoDistance off for two relatively distant points")
-	}	
-}
-
-func TestPathLengthInMeters(t *testing.T) {
-
-	l1 := PathLengthInMeters(path1)
-	if math.Abs(l1 - 2310) > 10 {
-		t.Errorf("path1 length calculated incorrectly")
-	}
-
-	l2 := PathLengthInMeters(path2)
-	if math.Abs(l2 - 1920) > 10 {
-		t.Errorf("path2 length calculated incorrectly")
-	}
-}
-
 func TestMakeGrid(t *testing.T) {
 
 	width := 7.0
@@ -283,26 +189,5 @@ func TestOverlaps(t *testing.T) {
 	r3 := PathBoundary(path3)
 	if r1.Overlaps(r3) != false {
 		t.Errorf("false positive error on overlap detection")
-	}
-}
-
-func TestPathSimilarity(t *testing.T) {
-
-	// test that the same path has a similarity of 1
-	// and two paths with different boundaries have a similarity of 0
-	// and two paths that are somewhat similar have a score between those two numbers
-	score := PathSimilarity(10.0, 10.0, path1, path1)
-	if score != 1 {
-		t.Errorf("the exact same path should return a score of 1")
-	}
-
-	score = PathSimilarity(10.0, 10.0, path1, path3)
-	if score != 0 {
-		t.Errorf("two paths with non-overlapping boundaries should return a score of 0")
-	}
-
-	score = PathSimilarity(10.0, 10.0, path1, path2)
-	if !(score > 0 && score < 1) {
-		t.Errorf("two somewhat similar paths should return a score between 0 and 1")
 	}
 }
